requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytz==2024.1
